- Squash bugs involving Decimal, String, Int conversions - Change colors in Quick Split view to match other views - Resize views so they are displayed correctly on all iPhone models compatible

+++ TipPoolCalc/EmployeeClass.swift
@@ -13,7 +13,7 @@ class Employee {
   var position: String = ""
   var weight: String = ""
   var hours: String = ""
-  var points: Int = 0
+  var points: Decimal = 0.00
   var tipsEarned: Decimal = 0.00
   
   init(name: String = "", position: String = "", weight: String = "", hours: String = "") {
